requirements.txt
aiohttp
ffmpeg-python
https://github.com/1Danish-00/Telethon/archive/master.zip
html_telegraph_poster
python-decouple
pyaesni
psutil
tgcrypto
